requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1
django-phonenumber-field==8.0.0
idna==3.10
inertia-django==0.6.0
phonenumbers==8.13.45
pillow==10.4.0
pytils==0.4.1
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.3